+++ src/nodes/abstractNode.tsx
@@ -1,14 +1,13 @@
+import React, { useEffect } from "react";
+import Draggable, { DraggableCore } from "react-draggable";
+
 export interface SinkDefinition {
   index: number;
   attr: string;
 }
 
 export interface NodeInputs {
-  [key: string]:
-    | undefined
-    | number
-    | string
-    | SinkDefinition;
+  [key: string]: undefined | number | string | SinkDefinition;
 }
 
 export interface NodeDefinition {
@@ -20,14 +19,128 @@ export interface NodeDefinition {
 type OutputValue = any;
 
 export interface NodeOutputs {
-    [key: string]: OutputValue;
-  }
-  
+  [key: string]: OutputValue;
+}
+
+export const VisualNode = (props: {
+  index: number;
+  title: string;
+  inputs: NodeInputs;
+  nodes: Array<BaseNode>;
+  setNodes: React.Dispatch<React.SetStateAction<NodeDefinition[]>>;
+}) => {
+  const node = props.nodes[props.index];
+  const getNodes = () => {
+    const newNodes = [...props.nodes.map((node) => node.getDefinition())];
+    return newNodes;
+  };
+
+  const handleChange = (key: string, value: string) => {
+    const newNodes = getNodes();
+    const node = newNodes[props.index];
+    node.inputs[key] = value;
+    props.setNodes(newNodes);
+  };
+
+  const handleMouseOver = () => {
+    const newNodes = getNodes();
+    const node: any = newNodes[props.index];
+    node.hilighted = true;
+    props.setNodes(newNodes);
+  };
+
+  const handleMouseOut = () => {
+    const newNodes = getNodes();
+    const node: any = newNodes[props.index];
+    node.hilighted = false;
+    props.setNodes(newNodes);
+  };
+
+  useEffect(() => {
+    // const setInput = async (callback: (args: any) => any) => {
+    //   const input = await nodeOutput(callback)
+    //   console.log(input())
+    // }
+    // setInput(testFunc)
+  }, []);
+
+  return (
+    <Draggable handle={`.handle`}>
+      <div
+        className="VisualNode"
+        style={{
+          // position: "absolute",
+          display: "inline-block",
+          // left: `${150 + (props.index * 5)}px`,
+          minHeight: "150px",
+          minWidth: "150px",
+          border: "1px solid black",
+          borderRadius: "5px",
+          background: "white",
+        }}
+        onMouseOver={() => handleMouseOver()}
+        onMouseOut={() => handleMouseOut()}
+      >
+        <span
+          className="handle"
+          id="VisualNode header"
+          style={{
+            display: "block",
+            borderBottom: "1px solid black",
+          }}
+        >
+          {props.title}
+        </span>
+        <div>
+          {Object.keys(props.inputs).map((key) => {
+            const [defaultValue, isFromSink] = node.getInputValue(key, props.nodes);
+            return (
+              <span
+                style={{
+                  display: "block",
+                }}
+              >
+                {key}:
+                {isFromSink ? (
+                  defaultValue
+                ) : (
+                  <input
+                    type={"text"}
+                    defaultValue={defaultValue}
+                    // onChange={(ev) => {
+                    //   setValue(ev.target.value);
+                    // }}
+                    //@ts-ignore
+                    onKeyUp={(ev) => handleChange(key, ev.target.value)}
+                  />
+                )}
+              </span>
+            );
+          })}
+        </div>
+        <div>
+          {Object.entries(node.outputs(props.nodes)).map(([key, value]) => {
+            return (
+              <span
+                style={{
+                  display: "block",
+                }}
+              >
+                {key}: {value}
+              </span>
+            );
+          })}
+        </div>
+      </div>
+    </Draggable>
+  );
+};
+
 export abstract class BaseNode {
   title: string;
   inputs: NodeInputs;
 
-  constructor({title, inputs}: {title: string, inputs: NodeInputs}) {
+  constructor({ title, inputs }: { title: string; inputs: NodeInputs }) {
     this.title = title;
     this.inputs = inputs;
   }
@@ -54,5 +167,26 @@ export abstract class BaseNode {
   }
 
   abstract outputs(nodes: BaseNode[]): NodeOutputs;
-  abstract getOutputValue(key: string, nodes: BaseNode[]): OutputValue;
+
+  getOutputValue(key: string, nodes: BaseNode[]) {
+    const outputs = this.outputs(nodes);
+
+    return outputs[key as keyof typeof outputs];
+  }
+
+  render(
+    index: number,
+    nodes: BaseNode[],
+    setNodes: React.Dispatch<React.SetStateAction<NodeDefinition[]>>
+  ) {
+    return (
+      <VisualNode
+        index={index}
+        title={this.title}
+        inputs={this.inputs}
+        nodes={nodes}
+        setNodes={setNodes}
+      />
+    );
+}
 }
