+++ src/nodes/htmlNode/htmlNode.tsx
@@ -1,12 +1,25 @@
-export interface HTMLNodeAttrs {
-    innerHTML?: string,
-    position?: string,
-    display?: string,
-    width?: number | string,
-    height?: number | string
-    top?: number | string,
-    left?: number | string,
-    right?: number | string,
-    bottom?: number | string
-};
+import { BaseNode, NodeInputs, SinkDefinition } from "../abstractNode";
 
+export class HtmlNode extends BaseNode {
+  inputs: {
+    html: string | SinkDefinition;
+  };
+
+  constructor({ title, inputs }: { title: string; inputs: NodeInputs }) {
+    super({ title, inputs });
+
+    if (typeof inputs.html === "object") {
+      this.inputs = {
+        html: inputs.html,
+      };
+    } else {
+      this.inputs = {
+        html: `${inputs.html || ""}`,
+      };
+    }
+  }
+
+  outputs() {
+    return {};
+  }
+}
